# engineering_docs/04_design/openapi-exam-pro-v1.yaml
...
openapi: 3.1.0
info:
  title: exam_pro API
  version: 1.3.0 # 修訂 2026-09-24（階段 5 整合回填）
  description: |
    家教專用數理題庫系統（exam_pro）之 HTTP API 契約。版本 v1.1，更新 2026-08-29，Owner：Ben（楊本顥）。
    本檔為端點清單唯一權威（對應 exam_pro/routes/index.js）；設計約定（錯誤格式、分頁、驗證規則）
    見 api_spec.md，文件不重複維護端點清單。全部 /api 路由經 apiKeyAuth（x-api-key，timing-safe）。
    對應需求：FR-001～FR-017。
    2026-08-29 修訂（PR #6/#7 程式碼同步）：新增 GET /chapter-volumes；listQuestions／generatePaper／createJob／QuestionCreate 補 source_type（FR-017）。
    2026-08-29 修訂之二（feat/source-detail 0007）：QuestionCreate／createJob 補 source_detail 來源註記；新增 POST /questions/batch-source 批次補標（FR-017 延伸）。
    2026-09-15f 修訂（feat/source-check 0009，FR-020）：GET /review 的 reason 值域加 transcription_mismatch（九個值）；job_questions.state 值域加 source_checked；GET /review/{jqId} 的 payload 可含 extract.source_text 與 source_check。端點形狀不變。
    2026-09-24 修訂（階段 5 教學診斷平台，FR-021～035；契約 docs/interfaces-stage5.md）：新增 GET /questions/{id}、/student-profile-options、
    /error-types、/kc、/kc/{id}、/questions/{id}/kcs、/students/{id}/weakness/kc、/students/{id}/remedial-paper、
    /students/{id}/remedial-paper/items、/coverage、/tutor、/voice/transcribe；擴充 PATCH /papers/{id}/results（錯因、部分給分、
    學生答案、註記；並更正 body 鍵名為 results）、GET /papers/{id}、GET /students/{id}/weakness（by_error_type）、學生三支（檔案欄位）、
    題目列表／新增／修改（文字詳解）、POST /generate-paper（blueprint）、POST /download-word（edition）、POST /jobs 與 GET /jobs/{id}（subject_group）。
...
    （回應多 incomplete_groups）。端點與 200 形狀不變。
servers:
  - url: http://localhost:3000/api
security:
  - ApiKeyAuth: []
...
paths:
...
  /students:
    get:
      tags: [students]
      summary: 學生清單（組卷下拉恆常需要，不掛旗標；裁決 S4-2）
      operationId: listStudents
      responses:
        "200": # 修訂 2026-09-24（階段 5 WS-A，FR-023）：每列在既有四欄之後多六個檔案欄位
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - type: object
                          properties:
                            id: { type: integer }
                            name: { type: string }
                            papers: { type: integer }
                            graded_ratio: { type: number }
                        - { $ref: "#/components/schemas/StudentProfile" }
    post:
      tags: [students]
      summary: 建立學生（可同時帶檔案欄位的任意子集；修訂 2026-09-24，FR-023）
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  required: [name]
                  properties:
                    name: { type: string, minLength: 1 }
                - { $ref: "#/components/schemas/StudentProfile" }
      responses:
        "201": { description: 已建立；回完整一列（id、name＋六個檔案欄位）, content: { application/json: { schema: { $ref: "#/components/schemas/StudentRow" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { description: 姓名重複 }
  /students/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    patch:
      tags: [students]
      summary: 學生改名／部分更新檔案（修訂 2026-09-24，FR-023：name 與六個檔案欄位的任意子集，沒送的不動；只送 name 的行為與訊息不變）
      operationId: renameStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: object
                  minProperties: 1
                  description: 一個可改的欄位都沒送 → 400「至少要提供一個要修改的欄位（name、grade、track、target_exams、school、textbook_version、note）。」
                  properties:
                    name: { type: string, minLength: 1 }
                - { $ref: "#/components/schemas/StudentProfile" }
      responses:
        "200": { description: 已更新；回完整一列, content: { application/json: { schema: { $ref: "#/components/schemas/StudentRow" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: 姓名重複 }
    delete:
      tags: [students]
      summary: 刪除學生
      operationId: deleteStudent
      responses:
        "200": { description: 已刪除 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /students/{id}/merge:
    post:
      tags: [students]
      summary: 學生併名（:id 併入 into_id；attempts 衝突保留目標側）
      operationId: mergeStudent
      parameters: [{ $ref: "#/components/parameters/IdPath" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [into_id]
              properties:
                into_id: { type: integer, minimum: 1, description: 併入目標學生 ID；不得等於來源 ID }
      responses:
        "200":
          description: 已合併
          content:
            application/json:
              schema:
                type: object
                properties:
                  moved_attempts: { type: integer }
                  dropped_conflicts: { type: integer, description: 目標已寫過同題而刪除的來源列數 }
                  moved_papers: { type: integer, description: 搬移至目標學生的試卷數（studentAdminController.mergeStudent） }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  schemas:
...
    StudentProfile:
      type: object
      description: 學生檔案欄位（白名單 config/studentProfile.js；FR-023）。POST／PATCH 可帶任意子集；PATCH 沒送的欄位不動
      properties:
        grade: { type: ["integer", "null"], enum: [10, 11, 12, null] }
        track: { type: ["string", "null"], enum: [自然組, 社會組, 未分組, null], description: 空字串＝null }
        target_exams: { type: ["array", "null"], uniqueItems: true, items: { type: string, enum: [學測, 分科, 統測, 段考, 其他] }, description: 'null 視同空陣列；存成白名單順序' }
        school: { type: ["string", "null"], maxLength: 50 }
        textbook_version: { type: ["string", "null"], enum: [龍騰, 翰林, 南一, 泰宇, 三民, 全華, 康熹, 其他, null] }
        note: { type: ["string", "null"], maxLength: 500 }
    StudentRow:
      description: POST／PATCH /students 的回應與 GET /students 的每列（GET 另有 papers、graded_ratio）
      allOf:
        - type: object
          properties:
            id: { type: integer }
            name: { type: string }
        - { $ref: "#/components/schemas/StudentProfile" }
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string, description: 人可讀訊息；正式環境不含內部錯誤細節（NFR-001） }
        errors: { type: array, items: { type: object }, description: 欄位級錯誤（僅驗證類 400 出現） }
  responses:
    BadRequest:
      description: 參數或欄位驗證失敗
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    NotFound:
      description: 資源不存在（含 FEATURE_* 旗標關閉時路由未掛載）
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: app.js 對 /api 全域套用之 apiKeyAuth；timing-safe 比對（NFR-001）